pyproject.toml
[project]
name = "terminal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.128.8",
    "httpx>=0.28.1",
    "oci>=2.167.1",
    "ocifs>=1.3.4",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn[standard]>=0.40.0",
    "cloudscraper>=1.2.71",
    "numpy>=2.4.2",
    "pyarrow>=23.0.0",
    "fastparquet>=2025.12.0",
    "pandas>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "orjson>=3.11.7",
    "websockets>=15.0",
    "protobuf>=6.33.5",
    "pydantic>=2.12.5",
    "cryptography>=44.0.0",
    "apscheduler>=3.11.2",
    "pywebpush>=2.0.1",
]



[tool.hatch.build.targets.wheel]
packages = ["src/terminal"]

[project.scripts]
terminal = "terminal.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "testcontainers[postgres]>=4.14.1",
    "pandas-stubs~=3.0.0",
    "grpcio-tools>=1.78.0",
]


[tool.pytest.ini_options]
asyncio_mode = "strict"
filterwarnings = [
    "ignore:'asyncio.iscoroutinefunction' is deprecated:DeprecationWarning:circuitbreaker",
]
